backend/requirements.txt
anyio==4.7.0
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.103.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pydantic==1.10.9
pyodbc==5.0.0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.1
watchfiles==1.0.3
websockets==14.1
